# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NLP Question Generator API",
    "version": "2.0.1",
    "description": "Multi-provider AI-powered API for generating multiple-choice quiz questions from text or files",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/api/generate": {
      "post": {
        "tags": ["Generation"],
        "summary": "Generate questions from text",
        "description": "Generate multiple-choice quiz questions from provided text using the current AI provider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "The text to generate questions from",
                    "minLength": 50,
                    "maxLength": 50000,
                    "example": "Python is a high-level programming language known for its simplicity and readability."
                  },
                  "numQuestions": {
                    "type": "integer",
                    "description": "Number of questions to generate",
                    "minimum": 1,
                    "maximum": 50,
                    "default": 10,
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Questions generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input"
          },
          "500": {
            "description": "Server error - generation failed"
          }
        }
      }
    },
...
    "/api/switch-provider": {
      "post": {
        "tags": ["Providers"],
        "summary": "Switch AI provider",
        "description": "Switch to a different AI provider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provider"],
                "properties": {
                  "provider": {
                    "type": "string",
                    "enum": [
                      "gemini",
                      "openai",
                      "anthropic",
                      "deepseek",
                      "kimi",
                      "kimicn"
                    ],
                    "description": "Provider name to switch to",
                    "example": "gemini"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider switched successfully"
          },
          "400": {
            "description": "Invalid provider or provider not configured"
          }
        }
      }
    },
...
    "/api/documents/inspect": {
      "post": {
        "tags": ["Documents"],
        "summary": "Inspect document content",
        "description": "Upload a document to extract structure (slides, images) for granular selection. Returns a docId for subsequent generation.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUpload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document inspected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInspectionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/generate": {
      "post": {
        "tags": ["Documents"],
        "summary": "Generate questions from selected content",
        "description": "Generate questions using a previously inspected document ID and a content filter.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["docId"],
                "properties": {
                  "docId": {
                    "type": "string",
                    "description": "The ID returned from /inspect",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "includeSlides": {
                        "type": "array",
                        "items": { "type": "integer" },
                        "description": "List of slide/page numbers to include (1-based)"
                      },
                      "includeImages": {
                        "type": "array",
                        "items": { "type": "string" },
                        "description": "List of image source identifiers to include"
                      },
                      "numQuestions": {
                        "type": "integer",
                        "default": 10
                      },
                      "difficulty": {
                        "type": "string",
                        "enum": ["easy", "medium", "hard", "mixed"]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Questions generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/parallel/estimate": {
      "get": {
        "tags": ["System"],
        "summary": "Estimate parallel savings",
        "description": "Estimate time savings for parallel processing based on question count",
        "parameters": [
          {
            "name": "numQuestions",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 30
            },
            "description": "Number of questions to estimate for"
          }
        ],
        "responses": {
          "200": {
            "description": "Estimate retrieved successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FileUpload": {
        "type": "object",
        "required": ["file"],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The file to upload"
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "questiontext": {
            "type": "string",
            "description": "The question text"
          },
          "optiona": {
            "type": "string",
            "description": "Option A"
          },
          "optionb": {
            "type": "string",
            "description": "Option B"
          },
          "optionc": {
            "type": "string",
            "description": "Option C"
          },
          "optiond": {
            "type": "string",
            "description": "Option D"
          },
          "correctanswer": {
            "type": "string",
            "enum": ["A", "B", "C", "D"],
            "description": "The correct answer"
          },
          "difficulty": {
            "type": "string",
            "enum": ["easy", "medium", "hard"],
            "description": "Question difficulty level"
          }
        }
      },
      "QuestionResponse": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          },
          "provider": {
            "type": "string",
            "description": "AI provider used"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "generated_at": {
                "type": "string",
                "format": "date-time"
              },
              "numQuestions": {
                "type": "integer"
              },
              "source": {
                "type": "string"
              },
              "model": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "DocumentInspectionResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "docId": {
            "type": "string",
            "description": "Unique identifier for the uploaded document state"
          },
          "filename": { "type": "string" },
          "pages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "page": { "type": "integer" },
                "text": { "type": "string" },
                "images": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "source": { "type": "string" },
                      "mediaType": { "type": "string" },
                      "data": {
                        "type": "string",
                        "description": "Base64 encoded image data"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}